+++ package.json
@@ -24,10 +24,10 @@
     "async": "0.9.0",
     "aws-sdk": "~2.1.8",
     "big-number": "0.3.1",
-    "body-parser": "1.10.2",
+    "body-parser": "1.11.0",
     "cookie-session": "~1.1.0",
-    "express": "~4.11.1",
-    "express-session": "~1.10.1",
+    "express": "~4.11.2",
+    "express-session": "~1.10.2",
     "http-auth": "~2.2.5",
     "jade": "~1.9.1",
     "kue": "~0.8.11",
@@ -36,7 +36,7 @@
     "passport": "0.2.1",
     "passport-steam": "0.1.4",
     "redis": "0.12.1",
-    "request": "~2.51.0",
+    "request": "~2.52.0",
     "winston": "0.8.3",
     "multer": "~0.1.7",
     "dotenv": "~0.5.1",
@@ -53,10 +53,10 @@
     "database-cleaner": "~0.9.1",
     "mongodb": "~1.4.29",
     "coveralls": "2.11.2",
-    "nock": "~0.59.0",
+    "nock": "~0.59.1",
     "zombie": "~2.5.1",
-    "watchify": "~2.2.1",
-    "browserify": "~8.1.1",
+    "watchify": "~2.3.0",
+    "browserify": "~8.1.3",
     "numeral": "~1.5.3",
     "datatables": "~1.10.5-dev",
     "jquery": "~2.1.3",
