+++ views/player_index.jade
@@ -83,4 +83,3 @@ block playercontent
   script.
     var data = !{JSON.stringify(player.histogramData)}
     generateHistograms(data);
-    generateCalHeatmap(data.calheatmap);
