+++ views/upload.jade
@@ -20,5 +20,3 @@ block content
   include ads/leaderboard_2
   
 block footer_assets
-  script.
-    uploadReplay();
